refactor app_settings, settings_local
